requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
commonmark==0.9.1
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
mypy-extensions==0.4.3
nhentei==2.0
pathspec==0.9.0
platformdirs==2.5.1
Pygments==2.11.2
PyYAML==6.0
requests==2.27.1
rich==12.1.0
textual==0.1.17
tomli==2.0.1
urllib3==1.26.9
yarl==1.7.2
